アバター画像設定コンポーネント<AvatarInput>を切り出す
