--- frontend/src/types.ts
@@ -1,5 +1,0 @@
-export type UserPersonalData = {
-  id: number;
-  email: string;
-  displayName: string;
-};
